+++ packages/core/src/typebox.ts
@@ -1,11 +1,16 @@
 import { type Hook, tbValidator } from "@hono/typebox-validator";
-import convert from "./toOpenAPISchema";
 import type { Static, TSchema } from "@sinclair/typebox";
 import { Value } from "@sinclair/typebox/value";
 import type { Env, MiddlewareHandler, ValidationTargets } from "hono";
-import type { OpenAPIRouteHandlerConfig, ResolverResult } from "./types";
-import { generateValidatorDocs, uniqueSymbol } from "./utils";
+import convert from "./toOpenAPISchema.js";
+import type { OpenAPIRouteHandlerConfig, ResolverResult } from "./types.js";
+import { generateValidatorDocs, uniqueSymbol } from "./utils.js";
 
+/**
+ * Generate a resolver for a TypeBox schema
+ * @param schema TypeBox schema
+ * @returns Resolver result
+ */
 export function resolver<T extends TSchema>(schema: T): ResolverResult {
   return {
     builder: async (options?: OpenAPIRouteHandlerConfig) => ({
@@ -17,6 +22,13 @@ export function resolver<T extends TSchema>(schema: T): ResolverResult {
   };
 }
 
+/**
+ * Create a validator middleware
+ * @param target Target for validation
+ * @param schema TypeBox schema
+ * @param hook Hook for validation
+ * @returns Middleware handler
+ */
 export function validator<
   T extends TSchema,
   Target extends keyof ValidationTargets,
@@ -25,11 +37,11 @@ export function validator<
   V extends {
     in: { [K in Target]: Static<T> };
     out: { [K in Target]: Static<T> };
-  }
+  },
 >(
   target: Target,
   schema: T,
-  hook?: Hook<Static<T>, E, P>
+  hook?: Hook<Static<T>, E, P>,
 ): MiddlewareHandler<E, P, V> {
   const middleware = tbValidator(target, schema, hook);